feat: create Header component with responsive navigation and menu functionality
